requirements.txt
fastapi
passlib
sqlalchemy
psycopg2-binary
uvicorn
PyJWT
pydantic[email]
bcrypt